♻️ Refactor recordset
